backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfminer.six
python-docx
reportlab
pydantic
python-dotenv
supabase
